Fixed maze exit

+++ Solution/Maze.h
@@ -117,8 +117,8 @@ private:
 	GLuint mazeVbo;
 	GLuint mazeVao;
 
-	const int scaleXZ = 2;
-	const int scaleY  = 2;
+	const float scaleXZ = 2;
+	const float scaleY  = 2;
 
 	// Private functions
 	void LoadTextures();
